Extract family member print components
